Add a playbook for installing and configuring Graphite.
This playbook will install carbon-cache and graphite-api to provide a
minimalistic starting point for metrics collection.

It currently installs all components from their upstream source (albeit
with defaults that point to tags, and not master/development branches).
It uses systemd, and so is tailored specifically to Ubuntu 16.04.

It currently does not support any of the advanced configuration such as
multiple carbon-cache process + carbon-relay.  It is also lacking many
variables to fully tweak the various configuration files, although
there's support for fully configuring the storage schemas.

There is a small "master" playbook -- edx-monitoring -- for running the
role, and in the future, supporting the culmination of other playbooks
to build out an all-in-one monitoring server: Graphite, Grafana, etc.