Make Graphite/Carbon more overridable.
We previously used a less-than-optimal way of storing the configuration
values for Graphite/Carbon, which meant you had to override the entire
configuration (including restating default values, etc) to change one
thing.

This now mirrors how we approach things in the Grafana role, where each
distinct value is a distinct variable that can be overridden separately,
and is pulled together at a higher level into the configuration.